Make >64k directory support actually work
Also remove some pointless global state. I have no idea if this
code had ever worked properly. It operated on the assumption the
end of central directory record is part of the central directory
and ended up requesting less data than needed since it omitted the
EOCD record size. The code was ugly anyway so I rewrote it with
slightly different logic, now we just explicitly check if the
start of the central directory is within the 64k chunk we've
requested and fetch the missing part if it's not. Should be
easier to understand than whatever was happening before.